pyproject.toml
[tool.poetry]
name = "fast-api-memo"
version = "0.1.0"
description = ""
authors = ["dakenoakari <dakeno-akari@natric.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
alembic = "^1.13.2"
pyjwt = "^2.8.0"
sqlalchemy = "^2.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
